requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Flask-Caching==2.1.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
APScheduler==3.10.4

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Web Scraping (for store APIs)
beautifulsoup4==4.12.2
selenium==4.15.2

# Email
Flask-Mail==0.9.1

# API Documentation
flask-restx==1.3.0

# Development
python-dateutil==2.8.2
Pillow==10.1.0
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Monitoring
structlog==23.2.0